+++ src/enemy.cpp
@@ -8,8 +8,8 @@ Enemy::Enemy(int spd, int lfe, int frm, int lftime, SDL_Rect src, SDL_Rect dst, 
 }
 
 void Enemy::UpdateHitbox() {
-    hitbox.w = rect_dst.w * 0.4;
-    hitbox.h = rect_dst.h * 0.85;
+    hitbox.w = static_cast<int>(rect_dst.w * 0.4);
+    hitbox.h = static_cast<int>(rect_dst.h * 0.85);
 
     hitbox.x = rect_dst.x + (rect_dst.w - hitbox.w) / 2 + 5;
     hitbox.y = rect_dst.y + (rect_dst.h - hitbox.h) / 2 + 10;
